requirements.txt
# Core Dependencies
openai>=1.12.0
python-dotenv>=1.0.0

# Data Management
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlite3  # Built-in with Python

# RAG Implementation
chromadb>=0.4.22
sentence-transformers>=2.3.1

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.2
